# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Volunteer Network API",
    "description": "A comprehensive API for connecting volunteers with organizations that need help",
    "version": "1.0.0",
    "contact": {
      "name": "Volunteer Network Support",
      "email": "support@volunteernetwork.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.volunteernetwork.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login or register endpoints"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier for the user"
          },
          "name": {
            "type": "string",
            "description": "User's full name",
            "minLength": 2,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "role": {
            "type": "string",
            "enum": ["volunteer", "organization"],
            "description": "User's role in the system"
          },
          "location": {
            "type": "string",
            "description": "User's location",
            "minLength": 2,
            "maxLength": 100
          },
          "bio": {
            "type": "string",
            "description": "User's biography",
            "maxLength": 500
          },
          "profileImage": {
            "type": "string",
            "description": "URL to user's profile image"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["name", "email", "role", "location"]
      },
      "Opportunity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier for the opportunity"
          },
          "title": {
            "type": "string",
            "description": "Title of the opportunity",
            "minLength": 5,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the opportunity",
            "minLength": 10,
            "maxLength": 1000
          },
          "location": {
            "type": "string",
            "description": "Location where the opportunity takes place",
            "minLength": 2,
            "maxLength": 100
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time of the opportunity"
          },
          "category": {
            "type": "string",
            "enum": ["education", "healthcare", "environment", "community", "animals", "other"],
            "description": "Category of the opportunity"
          },
          "duration": {
            "type": "string",
            "enum": ["1-2 hours", "3-5 hours", "6-8 hours", "Full day", "Multiple days"],
            "description": "Expected duration of the opportunity"
          },
          "maxVolunteers": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum number of volunteers needed"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the opportunity is currently active",
            "default": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/User"
          },
          "volunteersApplied": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "description": "List of volunteers who have applied"
          },
          "isFull": {
            "type": "boolean",
            "description": "Whether the opportunity has reached maximum volunteers"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["title", "description", "location", "date", "category", "duration"]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "description": "Current page number"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          },
          "totalDocs": {
            "type": "integer",
            "description": "Total number of documents"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether there is a next page"
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Whether there is a previous page"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field name that caused the error"
                },
                "message": {
                  "type": "string",
                  "description": "Specific error message for the field"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/opportunities": {
      "get": {
        "summary": "Get All Opportunities",
        "description": "Get all active opportunities with optional filtering and pagination",
        "tags": ["Opportunities"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Filter by location",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string",
              "enum": ["education", "healthcare", "environment", "community", "animals", "other"]
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Filter by date (opportunities after this date)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in title and description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunities retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "opportunities": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Opportunity"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Opportunity",
        "description": "Create a new volunteer opportunity (organizations only)",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Title of the opportunity",
                    "minLength": 5,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed description of the opportunity",
                    "minLength": 10,
                    "maxLength": 1000
                  },
                  "location": {
                    "type": "string",
                    "description": "Location where the opportunity takes place",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Date and time of the opportunity"
                  },
                  "category": {
                    "type": "string",
                    "enum": ["education", "healthcare", "environment", "community", "animals", "other"],
                    "description": "Category of the opportunity"
                  },
                  "duration": {
                    "type": "string",
                    "enum": ["1-2 hours", "3-5 hours", "6-8 hours", "Full day", "Multiple days"],
                    "description": "Expected duration of the opportunity"
                  },
                  "maxVolunteers": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Maximum number of volunteers needed"
                  }
                },
                "required": ["title", "description", "location", "date", "category", "duration"]
              },
              "example": {
                "title": "Help at Local Food Bank",
                "description": "We need volunteers to help sort and distribute food to families in need.",
                "location": "New York, NY",
                "date": "2024-02-15T10:00:00.000Z",
                "category": "community",
                "duration": "3-5 hours",
                "maxVolunteers": 10
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Opportunity created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Opportunity created successfully"
                    },
                    "opportunity": {
                      "$ref": "#/components/schemas/Opportunity"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Organization role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
} 